backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
greenlet>=3.0.0
asyncpg==0.30.0
aiosqlite==0.21.0
alembic==1.16.5
pydantic-settings==2.10.1
orjson==3.11.3
redis==6.4.0
httpx==0.28.1
pytest==8.4.2
pytest-asyncio==1.2.0
sortedcontainers==2.4.0
